Significant unittests updates
Created new unittests for testing all of the calculations performed in the manuscript, including checking that pandas dataframes produce the same expected results. Updated all jupyter notebooks in documentation by rerunning them and making sure they don't error. Added lines to automatically re-save csv files that get used as tables in the documentation. Updated the RunTests bash script to give the user the option to re-run all of the documentation jupyter notebooks. If there is an error, it will print to terminal but it will not stop the unit testing routine.